requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
click==8.0.3
Django==4.0
django-filter==21.1
djangorestframework==3.13.1
joblib==1.1.0
nltk==3.6.7
numpy==1.21.5
pandas==1.3.5
pickle5==0.0.11
pybind11==2.8.1
python-dateutil==2.8.2
pytz==2021.3
regex==2021.11.10
scikit-learn==0.24.1
scipy==1.7.3
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.0.0
tqdm==4.62.3
whitenoise==5.3.0
